feat(new-trace-view-trace-typ): moved tracewarnings to it's own file.
